requirements.txt
tkinter
pandas
python-docx
selenium
openpyxl 
